requirements.txt
openai-whisper
ffmpeg-python
flask
werkzeug
gunicorn
